+++ es6/expressions.d.ts
@@ -7,7 +7,12 @@ interface ParserOptions {
   literals?: { [x: string]: any };
   isIdentifierStart?: (char: string) => boolean;
   isIdentifierContinue?: (char: string) => boolean;
-  postEvaluate?: (result: any, tag: string, scope: any) => any;
+  postEvaluate?: (
+    result: any,
+    tag: string,
+    scope: any,
+    context: DXT.ParserContext
+  ) => any;
   evaluateIdentifier?: (
     tag: string,
     scope: any,