+++ es6/expressions.js
@@ -305,7 +305,7 @@ function configuredParser(config = {}) {
 					return "";
 				}
 				if (typeof config.postEvaluate === "function") {
-					result = config.postEvaluate(result, tag, scope);
+					result = config.postEvaluate(result, tag, scope, context);
 				}
 				return result;
 			},